add() method
Brief method that creates a new node and determines whether the list is empty or not. In the case the list is empty (but obviously exists) it makes n refer to both the head and the tail, and it has no previous element. Otherwise, it replaces the tail of the list.