addCompany implementation
Adds companies similar to addWatchlist in StockBroker, except it returns false in the if statement instead. It didn't appear to work when I did .put inside the if statement, so I changed it around. I also realised I didn't have a capital M in HashMap creation, so it was giving me an error