Had some troubles with add(index, obj)
Solved it by fixing some statement indentations and node pointers, as they were slightly off. The indentation initially caused an infinite loop, so the test did not pass/fail, just got stuck. Once that was fixed I found it was inserting at the wrong position, so needed to add -1 to index (so it traverses to the point before the desired position to perform the operations)